+++ src/main.c
@@ -113,7 +113,7 @@ parse_params (config_t *config, int argc, char *argv[])
           break;
         case 't':
           {
-            int number_of_cpus = sysconf (_SC_NPROCESSORS_ONLN);
+            long number_of_cpus = sysconf (_SC_NPROCESSORS_ONLN);
             config->number_of_threads = atoi (optarg);
             if (config->number_of_threads < 1
                 || config->number_of_threads > number_of_cpus)
