+++ src/thread_pool.c
@@ -281,7 +281,7 @@ thread_pool_join (thread_pool_t *thread_pool)
 }
 
 int
-create_threads (thread_pool_t *thread_pool, int number_of_threads,
+create_threads (thread_pool_t *thread_pool, long number_of_threads,
                 void *func (void *), void *context, size_t context_size,
                 char *name)
 {
